Added NoAmbiguousCharacters
Added Exact methods

NoAmbiguousCharacters removes LIOlio10 from the possible password pool

Exact allows for setting an exact number of a type of character.
ExactLower(5) means there will be exactly 5 lower case characters in the
password.  Note that calling With() or Require() after the fact will
remove the Exact logic and allow all characters of the type into the
pool.

Signed-off-by: Ken Herner <kherner@navistone.com>